pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "RAG System for Automated Task Review"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-genai>=1.36.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.0.8",
    "langchain-postgres>=0.0.12",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "mypy>=1.18.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.0",
    "pytest-timeout>=2.4.0",
    "pytest-xdist>=3.8.0",
    "ruff>=0.13.0",
]
